refactor: Move POINT_TYPES to data
