+++ src/data/pointTypes.json
@@ -1,0 +1,15 @@
+[
+  "island",
+  "route",
+  "cityboat",
+  "sauna",
+  "rent",
+  "sup",
+  "experience",
+  "charter",
+  "sightseeing",
+  "learn",
+  "taxi",
+  "visitor",
+  "station"
+]
